fix token requests

+++ src/actions/deleteUser.ts
@@ -1,10 +1,20 @@
 "use server"
 
 import { revalidateTag } from "next/cache"
+import { cookies } from "next/headers"
 type DeleteUserRequestType = string
 export async function deleteUser(id:DeleteUserRequestType){
+    const cookieStore = cookies()
+    const token = cookieStore.get("token")
+    if(token === undefined){
+        throw new Error("Token indefinido") 
+    }
     const response = await fetch(`${process.env.API_HOST}/users/${id}`, {
-        method: "DELETE"
+        method: "DELETE",
+        headers: {
+            'Content-Type': 'application/json',
+            'Authorization': `Bearer ${token}`,
+        },
     })
     const parsed = await response.json()
     if(response.status === 200){
